Changes to fix Open Watcom build warnings and errors. This should have been revision 7 which is unlikely to build.

+++ trunk/src/console.h
@@ -116,7 +116,7 @@ typedef struct {
     long X;
     long Y;
     unsigned short Buttons;
-    unsigned short Count;
+    unsigned long Count;
     TKeyCode KeyMask;
 } TMouseEvent;
